refactor(cmd): extract bundle/test-event cluster into endpoint_bundle.go
Continue the endpoint_diagnostics.go decomposition. Move the synthetic-event,
bundle, and validation helpers into a new endpoint_bundle.go: validationStage,
runEndpointTestEvent, runEndpointBundleDiagnostics, checkLogWritable,
stageFromCheck, redactLastEvent, redactConfig, writeJSONFile,
writeEventBundleFiles, syntheticEvent.

Move-only, same package: behavior-preserving. endpoint_diagnostics.go drops to
~530 lines; cmd tests pass.
